requirements.txt
streamlit
requests
Pillow
numpy
tensorflow
